requirements.txt
MetaTrader5
pandas
pyyaml
tensorflow
